Dropped in all changes made to support Digger III. Last of the prototype stuff.

+++ KAOSCommon/include/KAOS/Common/Utilities.h
@@ -42,7 +42,7 @@ namespace KAOS { namespace Common
 		{
 			str << std::setfill('0') << std::setw(width);
 		}
-		str << std::hex << value;
+		str << std::hex << static_cast<uint64_t>(value);
 
 		return str.str();
 	}
